Merge branch 'master' of github.com:benjyman/ben-astronomy

+++ moon/processing_scripts/namorrodor_magnus/cotter_moon.py
@@ -1,3 +1,4 @@
+#!/usr/bin/env python 
 #python script to cotter moon obs
 
 from astropy.io import fits
@@ -7,12 +8,18 @@ import cmd
 import datetime
 import numpy as np
 from casacore.tables import table,tablecolumn,tablerow
-
+import subprocess
 
 def cotter_moon(options):
-
+   machine=options.machine
+   if machine=='namorrodor':
+      mwa_dir = '/md0/moon/data/MWA/'
+   else:
+      mwa_dir = '/astro/mwaeor/MWA/data/'
+       
    obsid=options.obsid
-   sister_obsid=options.sister_obsid
+   if options.sister_obsid:
+      sister_obsid=options.sister_obsid
    
    time_res=(options.time_freq_res).split(',')[0]
    freq_res=(options.time_freq_res).split(',')[1]
@@ -22,9 +29,17 @@ def cotter_moon(options):
       track_off_moon_string= options.track_off_moon_list
    else:
       track_off_moon_string=' '
+
+   if options.no_dysco:
+      dysco_string=''
+   else:
+      dysco_string='-use-dysco'
    
-   print "cottering obsid %s paired with %s with time resolution %s and freq resolution %s" % (obsid,sister_obsid,time_res,freq_res)
-   
+   if options.sister_obsid:   
+      print "cottering obsid %s paired with %s with time resolution %s and freq resolution %s" % (obsid,sister_obsid,time_res,freq_res)
+   else:
+      print "cottering obsid %s with time resolution %s and freq resolution %s" % (obsid,time_res,freq_res)
+
    if options.track_off_moon:
       track_off_moon_list=track_off_moon_string.split(',')
 
@@ -60,27 +75,44 @@ def cotter_moon(options):
       off_moon_base_name="%s_%s_track_off_moon_paired_%s" % (sister_obsid,epoch_ID,obsid)
       off_moon_ms_name=mwa_dir+sister_obsid+'/'+off_moon_base_name+'.ms'
       ms_name=on_moon_ms_name
-   
-   if (options.track_off_moon):
+   elif (options.track_off_moon):
       on_moon_basename="%s_%s_trackmoon" % (sister_obsid,epoch_ID) 
       on_moon_ms_name=mwa_dir+sister_obsid+'/'+on_moon_basename+'.ms'
       off_moon_base_name=base_name+'_track_off_moon_paired_' + sister_obsid
       off_moon_ms_name=data_dir+off_moon_base_name+'.ms'
       ms_name=off_moon_ms_name
-      
+   else:
+      if (machine == "magnus" or machine == "galaxy"):
+         ms_name=base_name+'.ms'
+      else:
+         ms_name=data_dir+base_name+'.ms'
+
+   if epoch_ID[0:5]=="2015B":
+      #min size in Mb
+      min_ms_size=800.0
+   else:
+      min_ms_size=100.0
+
+   #Check if the ms already exists and is the right size then just exit - no need to make it again
+   ms_size = float(subprocess.check_output(['du','-sh', ms_name]).split()[0].decode('utf-8').split("'")[0].split('M')[0])
+   if (os.path.exists(ms_name) and ms_size >= min_ms_size):
+      print "%s already exists and has size %s (greatewr than min size %s), exiting cotter_moon.py" % (ms_name, ms_size, min_ms_size)
+      sys.exit(0)
+      #exit  
+
    #metafits_filename=data_dir+obsid+'.metafits'
    metafits_filename="%s%s_metafits_ppds.fits" % (data_dir,obsid)
+   
    
    flag_file_name=data_dir+obsid+'_flags.zip'
 
    #check if metafits file exists, if not, download it
-   if not os.path.isfile(metafits_filename):
-      #cmd="make_metafits.py -o %s -g %s" % (metafits_filename,obsid)
-      cmd="wget -O %s http://mwa-metadata01.pawsey.org.au/metadata/fits?obs_id=%s" % (metafits_filename,obsid)
-      print cmd
-      os.system(cmd)
-   else:
-      pass
+   #No always download new fits file
+   #if not os.path.isfile(metafits_filename):
+   #   #cmd="make_metafits.py -o %s -g %s" % (metafits_filename,obsid)
+   cmd="wget -O %s http://mwa-metadata01.pawsey.org.au/metadata/fits?obs_id=%s" % (metafits_filename,obsid)
+   print cmd
+   os.system(cmd)
 
    #unpack the flagfile
    cmd="unzip %s -d %s " % (flag_file_name,data_dir)
@@ -166,7 +198,7 @@ def cotter_moon(options):
    #need to put user options on the time and freq resolution - what is best for the new long baseline obs?
    #can probably get away with just halving each (double baselines)
    
-   cmd='cotter4 -flagfiles %s -norfi -o %s %s -m %s -use-dysco -timeres %s -freqres %s %s*gpubox*.fits' % (flagfiles_string,ms_name,track_moon_string,metafits_filename,time_res,freq_res,data_dir)
+   cmd='cotter4 -flagfiles %s -norfi -noantennapruning -o %s %s -m %s %s -timeres %s -freqres %s %s*gpubox*.fits' % (flagfiles_string,ms_name,track_moon_string,metafits_filename,dysco_string,time_res,freq_res,data_dir)
    print cmd
    os.system(cmd)
 
@@ -175,43 +207,44 @@ def cotter_moon(options):
       print flag_ants_cmd_string
       os.system(flag_ants_cmd_string)
    
-   if (options.cleanup and os.path.exists(ms_name)):
+   if (options.cleanup and os.path.exists(ms_name) and ms_size >= min_ms_size):
       cmd="rm -rf  %s*gpubox*.fits" % (data_dir)
       print cmd
       os.system(cmd)
 
+   #do all this stuff in the selfcal stage instead - so you are sure that both ms have already meen created
    #here merge the flag columns for both the on_moon and off_moon ms
    #only do this step for track_off_moon, (always need to make on_moon ms first)
-   if (options.track_off_moon):
-      on_moon_table=table(on_moon_ms_name,readonly=False)
-      #print on_moon_table
-      #on_moon_table_UVW=tablecolumn(on_moon_table,'UVW')
-      #on_moon_table_time=tablecolumn(on_moon_table,'TIME')
-      on_moon_table_flag=tablecolumn(on_moon_table,'FLAG')
-   
-      off_moon_table=table(off_moon_ms_name,readonly=False)
-      #print off_moon_table
-      #off_moon_table_UVW=tablecolumn(off_moon_table,'UVW')
-      #off_moon_table_time=tablecolumn(off_moon_table,'TIME')
-      off_moon_table_flag=tablecolumn(off_moon_table,'FLAG')
-   
-      new_off_moon_table_flag=np.logical_not(off_moon_table_flag)
-
-   
-      #look in obs_list_generator.py for how to compare LSTs
-   
-      #Not sure if 2018A obs are well enough LST matched
-      #Continue anyway (come back and look at 2015A) 
-   
-      #OR the two flag columns
-      new_flag_column=np.logical_or(on_moon_table_flag,off_moon_table_flag)
-      #print new_flag_column[20100:20101]
-   
-      on_moon_table.putcol('FLAG',new_flag_column)
-      off_moon_table.putcol('FLAG',new_flag_column)
-   
-      on_moon_table.close()
-      off_moon_table.close()
+   #if (options.track_off_moon):
+   #   on_moon_table=table(on_moon_ms_name,readonly=False)
+   #   #print on_moon_table
+   #   #on_moon_table_UVW=tablecolumn(on_moon_table,'UVW')
+   #   #on_moon_table_time=tablecolumn(on_moon_table,'TIME')
+   #   on_moon_table_flag=tablecolumn(on_moon_table,'FLAG')
+   #
+   #   off_moon_table=table(off_moon_ms_name,readonly=False)
+   #   #print off_moon_table
+   #   #off_moon_table_UVW=tablecolumn(off_moon_table,'UVW')
+   #   #off_moon_table_time=tablecolumn(off_moon_table,'TIME')
+   #   off_moon_table_flag=tablecolumn(off_moon_table,'FLAG')
+   #
+   #   new_off_moon_table_flag=np.logical_not(off_moon_table_flag)
+   #
+   # 
+   #   #look in obs_list_generator.py for how to compare LSTs
+   #
+   #   #Not sure if 2018A obs are well enough LST matched
+   #   #Continue anyway (come back and look at 2015A) 
+   #
+   #   #OR the two flag columns
+   #   new_flag_column=np.logical_or(on_moon_table_flag,off_moon_table_flag)
+   #   #print new_flag_column[20100:20101]
+   # 
+   #   on_moon_table.putcol('FLAG',new_flag_column)
+   #   off_moon_table.putcol('FLAG',new_flag_column)
+   #
+   #   on_moon_table.close()
+   #   off_moon_table.close()
 
    
 import sys,os
@@ -231,7 +264,8 @@ parser.add_option('--obsid',type='string', dest='obsid',default='',help='obsid t
 parser.add_option('--sister_obsid',type='string', dest='sister_obsid',default='',help='sister_obsid e.g. --sister_obsid="1199396880" [default=%default]')
 parser.add_option('--track_off_moon_list',type='string', dest='track_off_moon_list',default='',help='When track_off_moon is True. Details of on-moon pairing on_moon_obsid,RA,DEC of on_moon paired obs e.g. --track_off_moon_list="1199394880,13.0,0.56" [default=%default]')
 parser.add_option('--time_freq_res',type='string', dest='time_freq_res',default='8,80',help='Time and then frequency resolution, comma separated e.g. --time_freq_res="8,80" [default=%default]')
-
+parser.add_option('--machine',type='string', dest='machine',default='magnus',help='machine can be galaxy, magnus or namorrodor e.g. --machine="namorrodor" [default=%default]')
+parser.add_option('--no_dysco',action='store_true',dest='no_dysco',default=False,help='Do not use Dysco compression [default=%default]')
 
 
 (options, args) = parser.parse_args()
@@ -241,8 +275,6 @@ parser.add_option('--time_freq_res',type='string', dest='time_freq_res',default=
 #   track_off_moon_string= args[1]
 #else:
 #   track_off_moon_string=' '
-
-mwa_dir = '/md0/moon/data/MWA/'
 
 cotter_moon(options)
